Now use callflow.short with | as separator instead of spaces. callflow.short is going to have the following fields in the following order: time|trace|no|src|srcport|dir|dst|dstport|prot|summary|remark
time: is the 1st field, so it makes it easy to sort on time, and to merge
different callflow.short files with each other (especially useful if those are
obtained from non pcap files).

The 2nd field "trace", may hold the filename from which the data is retrieved.
This is more for future use.

The field dir (direction), may hold the sessionId.